+++ include/libpldm++/firmware_update.hpp
@@ -28,6 +28,7 @@ namespace fw_update
 		v1,
 		v1_1_0,
 		v1_2_0,
+		v1_3_0,
 	};
 
 	// forward declare structs and classes for our 'friend' declarations
@@ -69,6 +70,34 @@ namespace fw_update
 		// introduced in PackagePin::v1
 		const std::optional<std::string> vendorDefinedDescriptorTitle;
 		// introduced in PackagePin::v1
+		const std::vector<uint8_t> data;
+	};
+
+	struct ReferenceManifestData
+		: libpldm::GrowableStruct<struct ReferenceManifestData>,
+		  private libpldm::NonCopyableNonMoveable {
+	    private:
+		friend pldm::fw_update::PackageParser;
+
+		ReferenceManifestData(uint8_t SVHID,
+				      const std::vector<uint8_t> &vendorID,
+				      const std::vector<uint8_t> &data);
+
+	    public:
+		~ReferenceManifestData();
+
+		ReferenceManifestData(const ReferenceManifestData &ref);
+
+		bool operator==(const ReferenceManifestData &other) const;
+
+		// data members
+		// introduced in PackagePin::v1_3_0
+		const uint8_t SVHID;
+
+		// introduced in PackagePin::v1_3_0
+		const std::vector<uint8_t> vendorID;
+
+		// introduced in PackagePin::v1_3_0
 		const std::vector<uint8_t> data;
 	};
 
@@ -134,7 +163,9 @@ namespace fw_update
 			const std::string &componentImageSetVersionString,
 			const std::map<uint16_t, std::unique_ptr<DescriptorData> >
 				&descriptors,
-			const std::vector<uint8_t> &firmwareDevicePackageData);
+			const std::vector<uint8_t> &firmwareDevicePackageData,
+			const std::optional<ReferenceManifestData>
+				&referenceManifestData);
 
 	    public:
 		~FirmwareDeviceIDRecord();
@@ -169,6 +200,9 @@ namespace fw_update
 
 		// introduced in PackagePin::v1
 		const std::vector<uint8_t> firmwareDevicePackageData;
+
+		// introduced in PackagePin::v1_3_0
+		const std::optional<ReferenceManifestData> referenceManifestData;
 	};
 
 	struct DownstreamDeviceIDRecord
@@ -187,7 +221,9 @@ namespace fw_update
 			const std::map<uint16_t, std::unique_ptr<DescriptorData> >
 				&recordDescriptors,
 
-			const std::vector<uint8_t> &downstreamDevicePackageData);
+			const std::vector<uint8_t> &downstreamDevicePackageData,
+			const std::optional<ReferenceManifestData>
+				&downstreamDeviceReferenceManifestData);
 
 	    public:
 		~DownstreamDeviceIDRecord();
@@ -217,6 +253,10 @@ namespace fw_update
 
 		// introduced in PackagePin::v1_1_0
 		const std::vector<uint8_t> downstreamDevicePackageData;
+
+		// introduced in PackagePin::v1_3_0
+		const std::optional<ReferenceManifestData>
+			downstreamDeviceReferenceManifestData;
 	};
 
 	struct Package : libpldm::GrowableStruct<struct Package>,
@@ -292,6 +332,10 @@ namespace fw_update
 			struct pldm_package &package,
 			pldm_package_downstream_device_id_record
 				&downstreamDeviceId) noexcept;
+
+		static std::optional<pldm::fw_update::ReferenceManifestData>
+		getReferenceManifestData(
+			struct variable_field &reference_manifest_data) noexcept;
 	};
 
 } // namespace fw_update